Consolidate telecom snippets into a single module
* Merge dialer and voipapp sub-modules into a single telecom module
* Rename MainActivity to DialerActivity
* Rename VoipCallbackActivity to VoipCallActivity
* Update package structure and consolidate resources
* Update settings.gradle.kts to reflect module changes
