+++ settings.gradle.kts
@@ -50,6 +50,5 @@ include(
     ":contacts",
     ":cars",
     ":installprompt",
-    ":telecom:voipapp",
-    ":telecom:dialer"
+    ":telecom"
 )
