+++ telecom/.gitignore
@@ -1,0 +1,3 @@
+/build
+/src/androidTest/
+/src/test/
