--- telecom/dialer/.gitignore
@@ -1,2 +1,0 @@
-/build
-/src/androidTest/
