+++ telecom/src/main/AndroidManifest.xml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2026 The Android Open Source Project
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          https://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!--    [START android_telecom_own_calls_permission_entry]-->
+    <!-- Required permission for Jetpack Telecom -->
+    <uses-permission android:name="android.permission.MANAGE_OWN_CALLS" />
+    <!--    [END android_telecom_own_calls_permission_entry]-->
+
+    <application
+        android:allowBackup="true"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
+        android:roundIcon="@mipmap/ic_launcher_round"
+        android:supportsRtl="true"
+        android:theme="@style/Theme.Snippets">
+
+        <!--    [START android_telecom_activity_callback_handle_entry]-->
+        <!-- Activity to handle the callback intent from the system dialer -->
+        <activity
+            android:name=".VoipCallActivity"
+            android:exported="true">
+
+            <!-- Register callback intent -->
+            <intent-filter>
+                <action android:name="android.telecom.action.CALL_BACK" />
+            </intent-filter>
+        </activity>
+        <!--    [END android_telecom_activity_callback_handle_entry]-->
+
+        <activity
+            android:name=".DialerActivity"
+            android:exported="true"
+            android:theme="@style/Theme.Snippets" />
+    </application>
+
+</manifest>
