+++ telecom/src/main/res/mipmap-mdpi/ic_launcher.webp
